fix expire at

+++ src/model/token.go
@@ -4,5 +4,5 @@ type Token struct {
 	Key       string `json:"key"`
 	Data      string `json:"data"`
 	TouchedAt string `json:"touched_at"`
-	ExpireAt  int    `json:"expire_at"`
+	ExpireAt  int64  `json:"expire_at"`
 }
